+++ src/CDA.PCML/PCMLSample.cs
@@ -17,6 +17,7 @@ using System.Collections.Generic;
 using System.Xml;
 using CDA.Generator.Common.SCSModel.Interfaces;
 using Nehta.VendorLibrary.CDA;
+using Nehta.VendorLibrary.CDA.CDAModel;
 using Nehta.VendorLibrary.CDA.Common.Enums;
 using Nehta.VendorLibrary.CDA.Generator.Enums;
 using Nehta.VendorLibrary.CDA.SCSModel.Common;
@@ -276,28 +277,27 @@ namespace CDA.PCML
         /// <returns>PCML</returns>
         public static Nehta.VendorLibrary.CDA.Common.PCML PopulatePCML_1A(Boolean mandatorySectionsOnly)
         {
-            var healthCheckAssessment = Nehta.VendorLibrary.CDA.Common.PCML.CreatePCML();
+            var pharmacyCuratedMedsList = Nehta.VendorLibrary.CDA.Common.PCML.CreatePCML();
 
               // Include Logo
-            healthCheckAssessment.IncludeLogo = false;
+            pharmacyCuratedMedsList.IncludeLogo = false;
 
               // Set Creation Time
-            healthCheckAssessment.DocumentCreationTime = new ISO8601DateTime(DateTime.Now);
+            pharmacyCuratedMedsList.DocumentCreationTime = new ISO8601DateTime(DateTime.Now);
 
             #region Setup and populate the CDA context model
 
             // Setup and populate the CDA context model
-            var cdaContext = Nehta.VendorLibrary.CDA.Common.PCML.CreateCDAContext();
+            ICDAContextPCML cdaContext = Nehta.VendorLibrary.CDA.Common.PCML.CreateCDAContext();
+
             // Document Id
             cdaContext.DocumentId = BaseCDAModel.CreateIdentifier(BaseCDAModel.CreateOid(), null);
-
+            
             // Custodian
             cdaContext.Custodian = BaseCDAModel.CreateCustodian();
             GenericObjectReuseSample.HydrateCustodian(cdaContext.Custodian, mandatorySectionsOnly);
-            
 
             //Optional sections
-            // Legal authenticator
             if (!mandatorySectionsOnly)
             {
                 // Set Id  
@@ -305,41 +305,52 @@ namespace CDA.PCML
                 // CDA Context Version
                 cdaContext.Version = "1";
 
+                // Legal authenticator
                 cdaContext.LegalAuthenticator = BaseCDAModel.CreateLegalAuthenticator();
                 GenericObjectReuseSample.HydrateAuthenticator(cdaContext.LegalAuthenticator, mandatorySectionsOnly);
+
+                // Information Recipients
+                cdaContext.InformationRecipients = new List<IParticipationInformationRecipient>();
+
+                var recipient1 = BaseCDAModel.CreateInformationRecipient();
+                GenericObjectReuseSample.HydrateRecipient(recipient1, RecipientType.Primary, mandatorySectionsOnly);
+
+                var recipient2 = BaseCDAModel.CreateInformationRecipient();
+                GenericObjectReuseSample.HydrateRecipient(recipient2, RecipientType.Secondary, mandatorySectionsOnly);
+
+                cdaContext.InformationRecipients.AddRange(new[] { recipient1, recipient2 });
             }
 
-            healthCheckAssessment.CDAContext = cdaContext;
+
+            pharmacyCuratedMedsList.CDAContext = cdaContext;
             #endregion
 
+            
             #region Setup and Populate the SCS Context model
             // Setup and Populate the SCS Context model
 
-            healthCheckAssessment.SCSContext = Nehta.VendorLibrary.CDA.Common.PCML.CreateSCSContext();
+            pharmacyCuratedMedsList.SCSContext = Nehta.VendorLibrary.CDA.Common.PCML.CreateSCSContext();
 
             var authorHealthcareProvider = BaseCDAModel.CreateAuthorHealthcareProvider();
             GenericObjectReuseSample.HydrateAuthorHealthcareProvider(authorHealthcareProvider, mandatorySectionsOnly);
-            healthCheckAssessment.SCSContext.Author = authorHealthcareProvider;
+            pharmacyCuratedMedsList.SCSContext.Author = authorHealthcareProvider;
 
             //Cannot use as a device : prohibited by CORE Level One
-            //healthCheckAssessment.SCSContext.Author = GenericObjectReuseSample.CreateAuthorDevice();
+            //pharmacyCuratedMedsList.SCSContext.Author = GenericObjectReuseSample.CreateAuthorDevice();
 
-            healthCheckAssessment.SCSContext.Encounter = new Encounter
+            pharmacyCuratedMedsList.SCSContext.Encounter = new Encounter
             {
                 HealthcareFacility = PopulateHealthcareFacility(mandatorySectionsOnly)
             };
 
-            healthCheckAssessment.SCSContext.SubjectOfCare = BaseCDAModel.CreateSubjectOfCare();
-            GenericObjectReuseSample.HydrateSubjectofCare(healthCheckAssessment.SCSContext.SubjectOfCare, mandatorySectionsOnly);
+            pharmacyCuratedMedsList.SCSContext.SubjectOfCare = BaseCDAModel.CreateSubjectOfCare();
+            GenericObjectReuseSample.HydrateSubjectofCare(pharmacyCuratedMedsList.SCSContext.SubjectOfCare, mandatorySectionsOnly);
 
-
-           
 
             IParticipationPersonOrOrganisation person = Nehta.VendorLibrary.CDA.Common.PCML.CreateParticipationPersonOrOrganisation();
             person.Participant = Nehta.VendorLibrary.CDA.Common.PCML.CreateParticipantPersonOrOrganisation();
             person.Role = BaseCDAModel.CreateRole(HealthcareFacilityTypeCodes.AgedCareResidentialServices);
             person.Participant.Person = BaseCDAModel.CreatePersonWithOrganisation();
-
             
 
             var name1 = BaseCDAModel.CreatePersonName();
@@ -420,16 +431,16 @@ namespace CDA.PCML
             }
 
 
-            healthCheckAssessment.SCSContext.Participant = new List<IParticipationPersonOrOrganisation>();
-            healthCheckAssessment.SCSContext.Participant.Add(person);
+            pharmacyCuratedMedsList.SCSContext.Participant = new List<IParticipationPersonOrOrganisation>();
+            pharmacyCuratedMedsList.SCSContext.Participant.Add(person);
 
             #endregion
 
             #region Setup and populate the SCS Content model
             // Setup and populate the SCS Content model
-            healthCheckAssessment.SCSContent = Nehta.VendorLibrary.CDA.Common.PCML.CreateSCSContent();
+            pharmacyCuratedMedsList.SCSContent = Nehta.VendorLibrary.CDA.Common.PCML.CreateSCSContent();
 
-            healthCheckAssessment.SCSContent.EncapsulatedData = BaseCDAModel.CreateEncapsulatedData();
+            pharmacyCuratedMedsList.SCSContent.EncapsulatedData = BaseCDAModel.CreateEncapsulatedData();
 
             ExternalData report1 = EventSummary.CreateExternalData();
             report1.ExternalDataMediaType = MediaType.PDF;
@@ -438,10 +449,10 @@ namespace CDA.PCML
             
 
 
-            healthCheckAssessment.SCSContent.EncapsulatedData.ExternalData = report1;
+            pharmacyCuratedMedsList.SCSContent.EncapsulatedData.ExternalData = report1;
             #endregion
 
-            return healthCheckAssessment;
+            return pharmacyCuratedMedsList;
         }